pyproject.toml
[project]
name = "codeagent-basic"
version = "0.1.0"
description = ""
authors = [
    {name = "raistln",email = "samumarfon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub (>=0.29.2,<0.30.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "smolagents (>=1.10.0,<2.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "codeagent_basic", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
